pyproject.toml
[project]
name = "thirdeye"
dynamic = ["version"]
description = "Trace any agentic CLI to a unified local store, queryable from the CLI."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [ { name = "Duncan McKinnon" } ]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Operating System :: POSIX",
  "Operating System :: MacOS",
  "Topic :: Software Development",
  "Topic :: Utilities",
]
dependencies = [
  "click>=8.1",
  "msgpack>=1.0",
  "zstandard>=0.22",
  "pyaml>=23.0",
]

[project.urls]
Homepage = "https://github.com/duncankmckinnon/thirdeye"
Issues = "https://github.com/duncankmckinnon/thirdeye/issues"

[project.scripts]
thirdeye = "thirdeye.cli:main"
thirdeye-claude-session-start = "thirdeye.platforms.claude.hooks:session_start"
thirdeye-claude-user-prompt-submit = "thirdeye.platforms.claude.hooks:user_prompt_submit"
thirdeye-claude-pre-tool-use = "thirdeye.platforms.claude.hooks:pre_tool_use"
thirdeye-claude-post-tool-use = "thirdeye.platforms.claude.hooks:post_tool_use"
thirdeye-claude-stop = "thirdeye.platforms.claude.hooks:stop"
thirdeye-claude-subagent-stop = "thirdeye.platforms.claude.hooks:subagent_stop"
thirdeye-claude-stop-failure = "thirdeye.platforms.claude.hooks:stop_failure"
thirdeye-claude-notification = "thirdeye.platforms.claude.hooks:notification"
thirdeye-claude-permission-request = "thirdeye.platforms.claude.hooks:permission_request"
thirdeye-claude-session-end = "thirdeye.platforms.claude.hooks:session_end"

[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools_scm]
version_file = "src/thirdeye/_version.py"
fallback_version = "0.1.0.dev0"

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "pre-commit>=3.7",
  "ruff>=0.7",
  "build>=1.2",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = [
  "E501",   # line length handled by formatter
  "UP038",  # isinstance(x, (A, B)) is fine; X | Y is a stylistic preference
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
  "B007",   # unused loop control variable
  "B011",   # assert False
  "B017",   # blind exception assert
  "E741",   # ambiguous variable names (e.g. `l`)
  "F841",   # unused local (common in `with ... as x:` test patterns)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=thirdeye --cov-report=term-missing --cov-report=xml --cov-report=html"

[tool.coverage.run]
source = ["src/thirdeye"]
omit = ["src/thirdeye/_version.py"]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
